Mocks & New IRoboCommand for non-windows platforms (#217)
* Implement Mocks for consumer unit testing

* Create FactoryCommand and FactoryCommandFactory

- This new IRoboCommand utilizes the IFileCopierFactory to perform copy operations without creating a new RoboCopy process. This allows full control over the actual copy operation if desired, and allows for cross-platform targeting if needed.
- Added IAuthenticator to allow customization of the FactoryCommand's authentication prior to starting the command. By default simply verifies it can access source and destination.
- Adjusted helpers and extension methods as needed.

* Fix Tests not being async

* Rename FactoryCommand to RoboCommandPortable

- Since this new IRoboCommand is designed to be portable between platforms, the class name has been updated to reflect that.

* Update CopyFileExFactory.cs

* Implement RetryOptions

* Update Nuget Package for MSTest & Implement Testing Timeouts

* Move IFilePairExtensions and IDirectoryPairExtensions

* Resolve unit RoboCommandPortable tests (except Purge)

* Update Test App to allow testing new IRoboCommands

* WIP - Unit Testing

* Create Build-And-Test action

* Update build-and-test.yml

* Update RoboCommandPortable.cs

Claude Suggested Fixes

* Create CommandTests.cs

* Update CommandTests.cs

* Update CommandTests.cs

* Update ResultsBuilder.cs

- Provide a message to the log file if file logging is disabled via RoboSharpConfiguration

* Fix CommandTests.cs

Rework CommandTests.cs since Claude.ai was not producing passing tests. Tests are passing locally using RoboCommand.

* Update RoboCommandPortableTests.cs

Trigger RoboCommandPortable to test against the RoboCommand test suite

* Update ProgressEstimatorTests.cs

Fix for including 301_bytes file in output

* WIP - RoboCommandPortable

* CommandTests and Improved Documentation

* Update CommandTests.cs

* Consolidate & Resolve Tests cases & issues

- Consolidates redundant tests
- Fixes several tests for RoboCommandPortable
- CopyFileEx and StreamedCopier now remove the HIDDEN file attribute when overwriting files (otherwise files were marked threw 'unauthorized' exception)  -- achieves same functionality as RoboCopy

* Robosharp Unit Tests now use random destination with proper cleanup

* RoboCommandPortable is now passing all unit tests locally.

- Consolidated various test cases (for example, ones that were passing and tested for already in CommandTests<T>
- Tests should be using random directories now instead of a shared destination, which could potentially create conflicts.
- Removed TestPrep.cs since it is no longer relevant.
- Adjusted extensions as needed to get all tests passing and aligned with actual robocopy functionality.

* Update IFileCopierTests.cs

Increased dummy file size

* Refactor IFileCopierTests

- Testing granularity is improved and is now reusable more easily.

* Update Built-And-Test.yml

* Update build-and-test.yml

* Improve Test Logging for troubleshooting github action

* MockRoboCommandFactory exposed in namespace for clarity

* Update MockRoboCommandFactory.cs

* Update MainWindow.xaml

Add support for Logging paths to the MainWindow

* Revert change to RoboSharpConfig_EN

- RoboSharpConfig_EN had the 'Existing Directory' overridden, and some consumers rely on that functionality. Reverted back to previous value.
- Added unit tests for to ensure this does not occur in future.
   - This required isolating and exposing the function that transforms a log line into a ProcessedFileInfo for a directory or file. `RoboCommand.ParseFileorDirectory`

* Update RoboSharpConfiguration.cs

Transformed GetDefaultConfiguration() to use [GeneratedRegex]

* Revert change for default LogParsing_ExistingDir

* Update SymbolicLinkTests.cs

- Symbolic Link testing requires Admin privileges.
- Instead of declaring the test passed, it will not properly communicate that the test has been ignored if run without admin privileges.

* Update RoboCommand.cs

fixed typo

* Bump Version Number to 1.6.3.0

Normalize nuget package versions to 1.6.3.0

---------

+++ RoboSharp.Extensions.UnitTests/Windows/SymbolicLinkTests.cs
@@ -1,32 +1,63 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿#if WINDOWS
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RoboSharp.Extensions.SymbolicLinkSupport;
-using RoboSharp.UnitTests;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Windows.Win32.Storage.FileSystem;
+using System.Security.Principal;
 
 
 namespace RoboSharp.Extensions.Windows.UnitTests
 {
+    /// <summary>
+    /// Run the test only if running in with escalated privileges
+    /// </summary>
+    public sealed class RequiresAdminAttribute : ConditionBaseAttribute
+    {
+        public RequiresAdminAttribute() : base(ConditionMode.Include) { }
+
+        public override string GroupName => "Requires Admin Privileges";
+
+        public override bool IsConditionMet => IsAdministrator();
+
+        public static bool IsAdministrator()
+        {
+            using WindowsIdentity identity = WindowsIdentity.GetCurrent();
+            WindowsPrincipal principal = new WindowsPrincipal(identity);
+            bool result =  principal.IsInRole(WindowsBuiltInRole.Administrator);
+            return result;
+        }
+    }
+
     // These tests must be run as administrator to pass!
     [TestClass]
     public class SymbolicLinkTests
     {
         static readonly string Root = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "RoboSharp.SymbolicLinkTesting");
         static readonly string TargetRoot = Path.Combine(Root, "Targets");
+        static bool isAdmin = false;
 
         const string ERD = "\n>>\t";
         static string RandomName() => Path.GetFileNameWithoutExtension(Path.GetRandomFileName());
 
+        public TestContext TestContext { get; set;  }
+
+        [ClassInitialize]
+        public static void ClassInitialize(TestContext testContext)
+        {
+            isAdmin = RequiresAdminAttribute.IsAdministrator();
+        }
+
         [TestInitialize]
         public void Initialize_SymbolicLinkTest()
         {
+            if (!isAdmin)
+            {
+                Console.WriteLine("Test Requires Admin Privileges");
+                return;
+            }
             Console.WriteLine("----------------------\n THIS TEST REQUIRES ADMIN PRIVILEGES (OR DEVELOPER-MODE ENABLED) AT RUN-TIME ");
-            RoboSharp.UnitTests.Test_Setup.PrintEnvironment();
+            RoboSharp.UnitTests.Test_Setup.PrintEnvironment(TestContext);
             Directory.CreateDirectory(Root);
             Directory.CreateDirectory(TargetRoot);
             SymbolicLink.ALLOW_UNPRIVILEGED_CREATE = true;
@@ -35,10 +66,13 @@ namespace RoboSharp.Extensions.Windows.UnitTests
         [TestCleanup]
         public void Cleanup_SymbolicLinkTest()
         {
+            if (!isAdmin)
+                return;
             Directory.Delete(Root, true);
         }
 
         [TestMethod]
+        [RequiresAdmin]
         public void Test_SymbolicDirectory()
         {
             DirectoryInfo target = new DirectoryInfo(Path.Combine(TargetRoot, RandomName(), RandomName()));
@@ -97,6 +131,7 @@ namespace RoboSharp.Extensions.Windows.UnitTests
         }
 
         [TestMethod]
+        [RequiresAdmin]
         public void Test_SymbolicFile()
         {
             FileInfo target = new FileInfo(Path.Combine(TargetRoot, Path.GetFileNameWithoutExtension(Path.GetRandomFileName()) + ".txt"));
@@ -174,3 +209,5 @@ namespace RoboSharp.Extensions.Windows.UnitTests
 
     }
 }
+
+#endif
